refactor(review): extract ReviewContextLoader from ReviewOrchestrator (#250)
Step 3 of the decomposition: move the read side of the review pipeline
into a ReviewContextLoader that loads everything up front and returns a
ReviewContext record (diff, base comparison, prior reviews/comments,
persisted prior findings, instructions, repo labels, project stack, and
the first-visible / has-context signals).

Pure extract-class, no behaviour change — resolveMissingPrDetails,
fetchPrFiles, buildBaseComparison(+WithStats), buildDiffString,
fetchPriorReviews, botSummaryCommentExists, fetchIssueComments,
fetchPullRequestComments and resolveProjectStack move verbatim, and the
load() method mirrors the prior inline sequence exactly (#212, #74, #134
behaviours preserved). review() now calls contextLoader.load(...) and
reads the result; prClient, instructionsResolver and projectStackResolver
leave the orchestrator (deps drop below the S6539 threshold).

Tests: the read-method cases move to a new ReviewContextLoaderTest; the
review()-based resolve-missing-details integration tests stay in
ReviewOrchestratorTest (as ResolveMissingPrDetailsIntegration), which
wires a real ReviewContextLoader from the mocked clients. ReviewContext
defensively copies its list fields. Full suite green, spotbugs + spotless
clean.
