+++ src/test/java/dev/thiagogonzaga/thrillhousebot/review/ReviewOrchestratorTest.java
@@ -29,7 +29,6 @@ import dev.thiagogonzaga.thrillhousebot.github.*;
 import dev.thiagogonzaga.thrillhousebot.review.ai.AiReviewService;
 import dev.thiagogonzaga.thrillhousebot.review.ai.FindingVerificationService;
 import dev.thiagogonzaga.thrillhousebot.review.ai.ReviewResponse;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
@@ -154,10 +153,7 @@ class ReviewOrchestratorTest {
         authClient,
         reviewClient,
         commentClient,
-        prClient,
         reviewThreadService,
-        instructionsResolver,
-        projectStackResolver,
         aiReviewService,
         findingVerificationService,
         quoteValidator,
@@ -171,6 +167,17 @@ class ReviewOrchestratorTest {
         labeler,
         new CiStatusEvaluator(checkRunClient, prClient),
         new CheckRunManager(checkRunClient),
+        new ReviewContextLoader(
+            prClient,
+            reviewClient,
+            commentClient,
+            instructionsResolver,
+            projectStackResolver,
+            diffFormatter,
+            labeler,
+            followUpAnalyzer,
+            sessionPersistence,
+            config),
         objectMapper);
   }
 
@@ -179,280 +186,6 @@ class ReviewOrchestratorTest {
     // Plain concatenation: patches always use \n regardless of platform, so %n would be wrong
     var patch = "@@ -" + lineInPatch + ",1 +" + lineInPatch + ",1 @@\n-old\n+new";
     return new GitHubPullRequestClient.FileDiff(filename, "modified", 1, 1, 2, patch);
-  }
-
-  @Nested
-  class BuildDiffString {
-
-    @Test
-    void shouldReturnNoChangesForNullList() {
-      var result = orchestrator.buildDiffString(null);
-      assertEquals("(no changes detected)", result);
-    }
-
-    @Test
-    void shouldReturnNoChangesForEmptyList() {
-      var result = orchestrator.buildDiffString(Collections.emptyList());
-      assertEquals("(no changes detected)", result);
-    }
-
-    @Test
-    void shouldFormatSingleFileWithPatch() {
-      var files =
-          List.of(
-              new GitHubPullRequestClient.FileDiff(
-                  "src/main/Foo.java",
-                  "modified",
-                  5,
-                  3,
-                  8,
-                  "@@ -1,3 +1,5 @@\n unchanged\n+added line\n+another line"));
-
-      var result = orchestrator.buildDiffString(files);
-
-      assertTrue(result.contains("## Overview: 1 files (+5 -3)"));
-      assertTrue(result.contains("### src/main/Foo.java (modified, +5 -3)"));
-      assertTrue(result.contains("```diff"));
-      assertTrue(result.contains("+added line"));
-    }
-
-    @Test
-    void shouldFormatSingleFileWithoutPatch() {
-      var files =
-          List.of(
-              new GitHubPullRequestClient.FileDiff("binary-file.png", "modified", 0, 0, 0, null));
-
-      var result = orchestrator.buildDiffString(files);
-
-      assertTrue(result.contains("binary-file.png (modified, +0 -0)"));
-      assertFalse(result.contains("```diff"));
-    }
-
-    @Test
-    void shouldAccumulateTotalsAcrossMultipleFiles() {
-      var files =
-          List.of(
-              new GitHubPullRequestClient.FileDiff("a.java", "modified", 10, 2, 12, "@@ patch a"),
-              new GitHubPullRequestClient.FileDiff("b.java", "added", 25, 0, 25, "@@ patch b"),
-              new GitHubPullRequestClient.FileDiff("c.java", "modified", 3, 5, 8, null));
-
-      var result = orchestrator.buildDiffString(files);
-
-      assertTrue(result.contains("## Overview: 3 files (+38 -7)"));
-      assertTrue(result.contains("a.java"));
-      assertTrue(result.contains("b.java"));
-      assertTrue(result.contains("c.java"));
-    }
-
-    @Test
-    void shouldHandleFilesWithZeroChanges() {
-      var files =
-          List.of(
-              new GitHubPullRequestClient.FileDiff("renamed-only.txt", "renamed", 0, 0, 0, null));
-
-      var result = orchestrator.buildDiffString(files);
-
-      assertTrue(result.contains("## Overview: 1 files (+0 -0)"));
-      assertTrue(result.contains("renamed-only.txt (renamed, +0 -0)"));
-    }
-  }
-
-  @Nested
-  class BuildBaseComparison {
-
-    @Test
-    void shouldReturnSafeMessageForNullBase() {
-      var result = orchestrator.buildBaseComparison("auth", "owner", "repo", null, "abcdefgh");
-      assertEquals("(regression comparison unavailable — refs too short)", result);
-    }
-
-    @Test
-    void shouldReturnSafeMessageForNullHead() {
-      var result = orchestrator.buildBaseComparison("auth", "owner", "repo", "abcdefgh", null);
-      assertEquals("(regression comparison unavailable — refs too short)", result);
-    }
-
-    @Test
-    void shouldReturnSafeMessageForShortBaseSha() {
-      var result = orchestrator.buildBaseComparison("auth", "owner", "repo", "abc", "abcdefgh");
-      assertEquals("(regression comparison unavailable — refs too short)", result);
-    }
-
-    @Test
-    void shouldReturnSafeMessageForShortHeadSha() {
-      var result = orchestrator.buildBaseComparison("auth", "owner", "repo", "abcdefgh", "def");
-      assertEquals("(regression comparison unavailable — refs too short)", result);
-    }
-
-    @Test
-    void shouldReturnSafeMessageWhenBothNull() {
-      var result = orchestrator.buildBaseComparison("auth", "owner", "repo", null, null);
-      assertEquals("(regression comparison unavailable — refs too short)", result);
-    }
-
-    @Test
-    void shouldReturnNoChangesMessageWhenComparisonHasEmptyFiles() {
-      var emptyComparison = new GitHubPullRequestClient.CompareResponse(3, List.of());
-      when(prClient.compareCommits(
-              any(), any(), eq("owner"), eq("repo"), eq("abcdefgh"), eq("hijklmn")))
-          .thenReturn(emptyComparison);
-
-      var result =
-          orchestrator.buildBaseComparison("Bearer tok", "owner", "repo", "abcdefgh", "hijklmn");
-
-      assertEquals("(no changes between abcdefg and hijklmn)", result);
-    }
-
-    @Test
-    void shouldReturnNoChangesMessageWhenComparisonHasNullFiles() {
-      var nullFilesComparison = new GitHubPullRequestClient.CompareResponse(0, null);
-      when(prClient.compareCommits(
-              any(), any(), eq("owner"), eq("repo"), eq("abcdefgh"), eq("hijklmn")))
-          .thenReturn(nullFilesComparison);
-
-      var result =
-          orchestrator.buildBaseComparison("Bearer tok", "owner", "repo", "abcdefgh", "hijklmn");
-
-      assertEquals("(no changes between abcdefg and hijklmn)", result);
-    }
-
-    @Test
-    void shouldBuildComparisonWithFiles() {
-      var comparison =
-          new GitHubPullRequestClient.CompareResponse(
-              2,
-              List.of(
-                  new GitHubPullRequestClient.FileDiff(
-                      "src/Bar.java", "modified", 3, 1, 4, "@@ -1 +1,3 @@\n-old\n+new\n+extra")));
-      when(prClient.compareCommits(
-              any(), any(), eq("owner"), eq("repo"), eq("abcdefgh"), eq("hijklmn")))
-          .thenReturn(comparison);
-
-      var result =
-          orchestrator.buildBaseComparison("Bearer tok", "owner", "repo", "abcdefgh", "hijklmn");
-
-      assertTrue(result.contains("## Changes between base and head"));
-      assertTrue(result.contains("abcdefg..hijklmn: 2"));
-      assertTrue(result.contains("src/Bar.java"));
-      assertTrue(result.contains("```diff"));
-      assertTrue(result.contains("+new"));
-    }
-
-    @Test
-    void shouldSkipFilesWithoutPatchInComparison() {
-      var comparison =
-          new GitHubPullRequestClient.CompareResponse(
-              1,
-              List.of(
-                  new GitHubPullRequestClient.FileDiff("binary.bin", "modified", 0, 0, 0, null),
-                  new GitHubPullRequestClient.FileDiff(
-                      "src/Text.java", "modified", 1, 1, 2, "@@ patch")));
-      when(prClient.compareCommits(
-              any(), any(), eq("owner"), eq("repo"), eq("abcdefgh"), eq("hijklmn")))
-          .thenReturn(comparison);
-
-      var result =
-          orchestrator.buildBaseComparison("Bearer tok", "owner", "repo", "abcdefgh", "hijklmn");
-
-      assertTrue(result.contains("src/Text.java"));
-      assertFalse(result.contains("binary.bin"));
-    }
-
-    @Test
-    void shouldReturnUnavailableOnException() {
-      when(prClient.compareCommits(
-              any(), any(), eq("owner"), eq("repo"), eq("abcdefgh"), eq("hijklmn")))
-          .thenThrow(new RuntimeException("API down"));
-
-      var result =
-          orchestrator.buildBaseComparison("Bearer tok", "owner", "repo", "abcdefgh", "hijklmn");
-
-      assertEquals("(regression comparison unavailable)", result);
-    }
-  }
-
-  @Nested
-  class FetchPrFiles {
-
-    @Test
-    void shouldPropagateExceptionSoTheReviewFailsInsteadOfApproving() {
-      when(prClient.getPullRequestFiles(any(), any(), eq("owner"), eq("repo"), eq(42)))
-          .thenThrow(new RuntimeException("GitHub API error"));
-
-      // Swallowing the failure into an empty list would yield an empty diff and a false APPROVE on
-      // unreviewed code; it must propagate so review() takes the failure path.
-      assertThrows(
-          RuntimeException.class, () -> orchestrator.fetchPrFiles("auth", "owner", "repo", 42));
-    }
-
-    @Test
-    void shouldPropagateNotAuthorizedRatherThanReturnEmpty() {
-      when(prClient.getPullRequestFiles(any(), any(), eq("owner"), eq("repo"), eq(123)))
-          .thenThrow(new jakarta.ws.rs.NotAuthorizedException("Bad credentials"));
-
-      assertThrows(
-          jakarta.ws.rs.NotAuthorizedException.class,
-          () -> orchestrator.fetchPrFiles("auth", "owner", "repo", 123));
-    }
-
-    @Test
-    void shouldReturnFilesOnSuccess() {
-      var expected =
-          List.of(
-              new GitHubPullRequestClient.FileDiff("README.md", "modified", 2, 1, 3, "@@ patch"));
-      when(prClient.getPullRequestFiles(any(), any(), eq("owner"), eq("repo"), eq(1)))
-          .thenReturn(expected);
-
-      var result = orchestrator.fetchPrFiles("auth", "owner", "repo", 1);
-
-      assertEquals(expected, result);
-      assertEquals(1, result.size());
-    }
-  }
-
-  @Nested
-  class FetchPriorReviews {
-
-    @Test
-    void shouldReturnEmptyListOnException() {
-      when(reviewClient.listReviews(any(), any(), eq("owner"), eq("repo"), eq(42)))
-          .thenThrow(new RuntimeException("GitHub API error"));
-
-      var result = orchestrator.fetchPriorReviews("auth", "owner", "repo", 42);
-
-      assertNotNull(result);
-      assertTrue(result.isEmpty());
-    }
-
-    @Test
-    void shouldReturnEmptyListOnNotFound() {
-      when(reviewClient.listReviews(any(), any(), eq("owner"), eq("repo"), eq(99)))
-          .thenThrow(new jakarta.ws.rs.NotFoundException("Not found"));
-
-      var result = orchestrator.fetchPriorReviews("auth", "owner", "repo", 99);
-
-      assertNotNull(result);
-      assertTrue(result.isEmpty());
-    }
-
-    @Test
-    void shouldReturnReviewsOnSuccess() {
-      var expected =
-          List.of(
-              new GitHubReviewClient.ReviewResponse(
-                  1L,
-                  "Looks good",
-                  "APPROVED",
-                  "abc123",
-                  new GitHubReviewClient.ReviewResponse.User("some-user")));
-      when(reviewClient.listReviews(any(), any(), eq("owner"), eq("repo"), eq(10)))
-          .thenReturn(expected);
-
-      var result = orchestrator.fetchPriorReviews("auth", "owner", "repo", 10);
-
-      assertEquals(1, result.size());
-      assertEquals("Looks good", result.get(0).body());
-    }
   }
 
   @Nested
@@ -3062,205 +2795,6 @@ class ReviewOrchestratorTest {
   }
 
   @Nested
-  class ResolveMissingPrDetails {
-
-    private ReviewOrchestrator.ReviewRequest manualRequest(String sha) {
-      return new ReviewOrchestrator.ReviewRequest(
-          "owner", "repo", 7, sha, "(manual review)", "", "", "main", 123L, true);
-    }
-
-    @Test
-    void shouldFetchPrDetailsWhenShaIsBlank() {
-      when(prClient.getPullRequest(anyString(), anyString(), eq("owner"), eq("repo"), eq(7)))
-          .thenReturn(
-              new GitHubPullRequestClient.PullRequestDetails(
-                  "add new API",
-                  "Adds a new API endpoint",
-                  new GitHubPullRequestClient.Ref("headsha1234"),
-                  new GitHubPullRequestClient.Ref("basesha5678")));
-
-      var resolved = orchestrator.resolveMissingPrDetails("Bearer tok", manualRequest(""));
-
-      assertEquals("headsha1234", resolved.commitSha());
-      assertEquals("basesha5678", resolved.baseSha());
-      assertEquals("add new API", resolved.prTitle());
-      assertEquals("Adds a new API endpoint", resolved.prDescription());
-      assertTrue(resolved.isManualTrigger());
-    }
-
-    @Test
-    void shouldNotFetchWhenShaIsPresent() {
-      var req = manualRequest("abcdefgh");
-
-      var resolved = orchestrator.resolveMissingPrDetails("Bearer tok", req);
-
-      assertSame(req, resolved);
-      verifyNoInteractions(prClient);
-    }
-
-    @Test
-    void shouldTolerateMissingRefsInPrDetails() {
-      when(prClient.getPullRequest(anyString(), anyString(), eq("owner"), eq("repo"), eq(7)))
-          .thenReturn(new GitHubPullRequestClient.PullRequestDetails(null, null, null, null));
-
-      var resolved = orchestrator.resolveMissingPrDetails("Bearer tok", manualRequest(null));
-
-      assertEquals("", resolved.commitSha());
-      assertEquals("", resolved.baseSha());
-      assertEquals("(manual review)", resolved.prTitle());
-      assertEquals("", resolved.prDescription());
-    }
-
-    @Test
-    void reviewShouldPostFailureCommentWhenPrResolutionThrows() {
-      try (var mockedStatic = mockStatic(ReviewSession.class)) {
-        var session = mock(ReviewSession.class);
-        session.id = 1L;
-        when(session.getRepository()).thenReturn("owner/repo");
-        when(session.getPrNumber()).thenReturn(7);
-        when(session.getPrTitle()).thenReturn("(manual review)");
-        when(session.getCommitSha()).thenReturn("");
-        when(session.getTimestamp()).thenReturn(java.time.Instant.parse("2025-06-01T12:00:00Z"));
-        mockedStatic
-            .when(() -> ReviewSession.create(anyString(), anyInt(), anyString(), anyString()))
-            .thenReturn(session);
-
-        when(authClient.getAuthHeader(123L)).thenReturn("Bearer test");
-        when(prClient.getPullRequest(anyString(), anyString(), eq("owner"), eq("repo"), eq(7)))
-            .thenThrow(new RuntimeException("PR was deleted"));
-
-        orchestrator.review(manualRequest(""));
-
-        // A failing PR fetch takes the standard failure path instead of escaping to the
-        // dispatcher: failure comment, failed session, broadcast — and no check run calls
-        verify(commentClient)
-            .createComment(
-                anyString(),
-                anyString(),
-                eq("owner"),
-                eq("repo"),
-                eq(7),
-                argThat(req -> req.body().contains("could not be completed")));
-        verify(session).setStatus(ReviewSession.STATUS_FAILED);
-        verify(broadcaster, times(2)).broadcast(any(SessionEventBroadcaster.SessionEvent.class));
-        verify(checkRunClient, never())
-            .createCheckRun(anyString(), anyString(), anyString(), anyString(), any());
-      }
-    }
-
-    @Test
-    void reviewShouldUpdateSessionWithResolvedTitleAndSha() {
-      try (var mockedStatic = mockStatic(ReviewSession.class)) {
-        var session = mock(ReviewSession.class);
-        session.id = 1L;
-        when(session.getRepository()).thenReturn("owner/repo");
-        when(session.getPrNumber()).thenReturn(7);
-        when(session.getPrTitle()).thenReturn("(manual review)");
-        when(session.getCommitSha()).thenReturn("");
-        when(session.getTimestamp()).thenReturn(java.time.Instant.parse("2025-06-01T12:00:00Z"));
-        mockedStatic
-            .when(() -> ReviewSession.create(anyString(), anyInt(), anyString(), anyString()))
-            .thenReturn(session);
-
-        when(authClient.getAuthHeader(123L)).thenReturn("Bearer test");
-        when(prClient.getPullRequest(any(), any(), eq("owner"), eq("repo"), eq(7)))
-            .thenReturn(
-                new GitHubPullRequestClient.PullRequestDetails(
-                    "add new API",
-                    "Adds a new API endpoint",
-                    new GitHubPullRequestClient.Ref("headsha1234", "feature"),
-                    new GitHubPullRequestClient.Ref("basesha5678", "main")));
-        when(checkRunClient.createCheckRun(
-                anyString(), anyString(), anyString(), anyString(), any()))
-            .thenReturn(new GitHubCheckRunClient.CheckRunResponse(1L, "http://check"));
-        when(prClient.getPullRequestFiles(
-                anyString(), anyString(), anyString(), anyString(), anyInt()))
-            .thenReturn(List.of());
-        when(checkRunClient.getRequiredStatusChecks(
-                anyString(), anyString(), anyString(), anyString(), anyString()))
-            .thenReturn(
-                new GitHubCheckRunClient.RequiredStatusChecks(
-                    List.of("required-build"), List.of()));
-        when(prClient.compareCommits(
-                anyString(), anyString(), anyString(), anyString(), anyString(), anyString()))
-            .thenReturn(new GitHubPullRequestClient.CompareResponse(0, List.of()));
-        when(reviewClient.listReviews(anyString(), anyString(), anyString(), anyString(), anyInt()))
-            .thenReturn(List.of());
-        when(instructionsResolver.resolve(anyString(), anyString(), anyString(), anyLong()))
-            .thenReturn(InstructionsResolver.ResolvedInstructions.EMPTY);
-        when(aiReviewService.review(any(ReviewSession.class), any()))
-            .thenReturn(new ReviewResponse(List.of(), List.of(), null));
-
-        orchestrator.review(manualRequest(""));
-
-        // The session is created with the placeholder, then updated with resolved PR details
-        verify(session).setPrTitle("add new API");
-        verify(session).setCommitSha("headsha1234");
-        var checkRunCaptor =
-            ArgumentCaptor.forClass(GitHubCheckRunClient.CreateCheckRunRequest.class);
-        verify(checkRunClient)
-            .createCheckRun(
-                anyString(), anyString(), eq("owner"), eq("repo"), checkRunCaptor.capture());
-        assertEquals("headsha1234", checkRunCaptor.getValue().headSha());
-      }
-    }
-
-    @Test
-    void reviewShouldCatchExceptionWhenFetchingStatusChecks() {
-      try (var mockedStatic = mockStatic(ReviewSession.class)) {
-        var session = mock(ReviewSession.class);
-        session.id = 1L;
-        when(session.getRepository()).thenReturn("owner/repo");
-        when(session.getPrNumber()).thenReturn(7);
-        when(session.getPrTitle()).thenReturn("(manual review)");
-        when(session.getCommitSha()).thenReturn("");
-        when(session.getTimestamp()).thenReturn(java.time.Instant.parse("2025-06-01T12:00:00Z"));
-        mockedStatic
-            .when(() -> ReviewSession.create(anyString(), anyInt(), anyString(), anyString()))
-            .thenReturn(session);
-
-        when(authClient.getAuthHeader(123L)).thenReturn("Bearer test");
-        when(prClient.getPullRequest(any(), any(), eq("owner"), eq("repo"), eq(7)))
-            .thenReturn(
-                new GitHubPullRequestClient.PullRequestDetails(
-                    "add new API",
-                    "Adds a new API endpoint",
-                    new GitHubPullRequestClient.Ref("headsha1234", "feature"),
-                    new GitHubPullRequestClient.Ref("basesha5678", "main")));
-        when(prClient.getPullRequestFiles(
-                anyString(), anyString(), anyString(), anyString(), anyInt()))
-            .thenReturn(List.of());
-        when(checkRunClient.getRequiredStatusChecks(
-                anyString(), anyString(), anyString(), anyString(), anyString()))
-            .thenThrow(new RuntimeException("failed"));
-        when(checkRunClient.createCheckRun(
-                anyString(), anyString(), anyString(), anyString(), any()))
-            .thenReturn(new GitHubCheckRunClient.CheckRunResponse(1L, "http://check"));
-        when(prClient.compareCommits(
-                anyString(), anyString(), anyString(), anyString(), anyString(), anyString()))
-            .thenReturn(new GitHubPullRequestClient.CompareResponse(0, List.of()));
-        when(reviewClient.listReviews(anyString(), anyString(), anyString(), anyString(), anyInt()))
-            .thenReturn(List.of());
-        when(instructionsResolver.resolve(anyString(), anyString(), anyString(), anyLong()))
-            .thenReturn(InstructionsResolver.ResolvedInstructions.EMPTY);
-        when(aiReviewService.review(any(ReviewSession.class), any()))
-            .thenReturn(new ReviewResponse(List.of(), List.of(), null));
-
-        orchestrator.review(
-            new ReviewOrchestrator.ReviewRequest(
-                "owner", "repo", 7, "", "(manual review)", "", "", "main", 123L, true));
-
-        // The thrown getRequiredStatusChecks must be swallowed: the review still finalizes its
-        // check run and posts a review rather than failing the whole run.
-        verify(checkRunClient)
-            .updateCheckRun(anyString(), anyString(), anyString(), anyString(), anyLong(), any());
-        verify(reviewClient)
-            .createReview(anyString(), anyString(), anyString(), anyString(), anyInt(), any());
-      }
-    }
-  }
-
-  @Nested
   class UnresolvedPreviousGating {
 
     private final FollowUpAnalyzer realAnalyzer = new FollowUpAnalyzer(new ObjectMapper());
@@ -3764,135 +3298,159 @@ class ReviewOrchestratorTest {
   }
 
   @Nested
-  class FetchPullRequestComments {
+  class ResolveMissingPrDetailsIntegration {
 
-    @Test
-    void shouldReturnEmptyListWhenClientThrows() {
-      // The paginating client never returns null, so the only failure to absorb is the fetch
-      // throwing — it degrades to an empty list rather than aborting the review.
-      when(reviewClient.listPullRequestComments(
-              anyString(), anyString(), anyString(), anyString(), anyInt()))
-          .thenThrow(new RuntimeException("boom"));
-
-      assertTrue(orchestrator.fetchPullRequestComments("auth", "owner", "repo", 1).isEmpty());
+    private ReviewOrchestrator.ReviewRequest manualRequest(String sha) {
+      return new ReviewOrchestrator.ReviewRequest(
+          "owner", "repo", 7, sha, "(manual review)", "", "", "main", 123L, true);
     }
 
     @Test
-    void shouldReturnCommentsFromClient() {
-      var comment =
-          new GitHubReviewClient.PullRequestComment(
-              1L, null, "f", "body", new GitHubReviewClient.ReviewResponse.User("u"));
-      when(reviewClient.listPullRequestComments(
-              anyString(), anyString(), anyString(), anyString(), anyInt()))
-          .thenReturn(List.of(comment));
+    void reviewShouldPostFailureCommentWhenPrResolutionThrows() {
+      try (var mockedStatic = mockStatic(ReviewSession.class)) {
+        var session = mock(ReviewSession.class);
+        session.id = 1L;
+        when(session.getRepository()).thenReturn("owner/repo");
+        when(session.getPrNumber()).thenReturn(7);
+        when(session.getPrTitle()).thenReturn("(manual review)");
+        when(session.getCommitSha()).thenReturn("");
+        when(session.getTimestamp()).thenReturn(java.time.Instant.parse("2025-06-01T12:00:00Z"));
+        mockedStatic
+            .when(() -> ReviewSession.create(anyString(), anyInt(), anyString(), anyString()))
+            .thenReturn(session);
 
-      assertEquals(
-          List.of(comment), orchestrator.fetchPullRequestComments("auth", "owner", "repo", 1));
-    }
-  }
+        when(authClient.getAuthHeader(123L)).thenReturn("Bearer test");
+        when(prClient.getPullRequest(anyString(), anyString(), eq("owner"), eq("repo"), eq(7)))
+            .thenThrow(new RuntimeException("PR was deleted"));
 
-  @Nested
-  class BotSummaryCommentDetection {
+        orchestrator.review(manualRequest(""));
 
-    private GitHubCommentClient.IssueComment comment(String body, String login) {
-      return new GitHubCommentClient.IssueComment(
-          body, login == null ? null : new GitHubReviewClient.ReviewResponse.User(login));
-    }
-
-    private void stubComments(GitHubCommentClient.IssueComment... comments) {
-      when(commentClient.listComments(anyString(), anyString(), anyString(), anyString(), anyInt()))
-          .thenReturn(List.of(comments));
-    }
-
-    @Test
-    void shouldDetectABotSummaryComment() {
-      stubComments(
-          comment("@thrillhousebot please review", "someuser"),
-          comment(PrSummaryGenerator.SUMMARY_HEADING + "\n\nbody", "thrillhousebot[bot]"));
-
-      assertTrue(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
-    }
-
-    @Test
-    void shouldTolerateLeadingWhitespaceBeforeTheHeading() {
-      stubComments(
-          comment("\n  " + PrSummaryGenerator.SUMMARY_HEADING + "\n", "thrillhousebot[bot]"));
-
-      assertTrue(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
+        // A failing PR fetch takes the standard failure path instead of escaping to the
+        // dispatcher: failure comment, failed session, broadcast — and no check run calls
+        verify(commentClient)
+            .createComment(
+                anyString(),
+                anyString(),
+                eq("owner"),
+                eq("repo"),
+                eq(7),
+                argThat(req -> req.body().contains("could not be completed")));
+        verify(session).setStatus(ReviewSession.STATUS_FAILED);
+        verify(broadcaster, times(2)).broadcast(any(SessionEventBroadcaster.SessionEvent.class));
+        verify(checkRunClient, never())
+            .createCheckRun(anyString(), anyString(), anyString(), anyString(), any());
+      }
     }
 
     @Test
-    void shouldIgnoreASummaryHeadingPostedBySomeoneElse() {
-      // A human (or another bot) quoting the heading must not be mistaken for the bot's own
-      // summary.
-      stubComments(comment(PrSummaryGenerator.SUMMARY_HEADING + "\n\nspoof", "impersonator"));
+    void reviewShouldUpdateSessionWithResolvedTitleAndSha() {
+      try (var mockedStatic = mockStatic(ReviewSession.class)) {
+        var session = mock(ReviewSession.class);
+        session.id = 1L;
+        when(session.getRepository()).thenReturn("owner/repo");
+        when(session.getPrNumber()).thenReturn(7);
+        when(session.getPrTitle()).thenReturn("(manual review)");
+        when(session.getCommitSha()).thenReturn("");
+        when(session.getTimestamp()).thenReturn(java.time.Instant.parse("2025-06-01T12:00:00Z"));
+        mockedStatic
+            .when(() -> ReviewSession.create(anyString(), anyInt(), anyString(), anyString()))
+            .thenReturn(session);
 
-      assertFalse(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
+        when(authClient.getAuthHeader(123L)).thenReturn("Bearer test");
+        when(prClient.getPullRequest(any(), any(), eq("owner"), eq("repo"), eq(7)))
+            .thenReturn(
+                new GitHubPullRequestClient.PullRequestDetails(
+                    "add new API",
+                    "Adds a new API endpoint",
+                    new GitHubPullRequestClient.Ref("headsha1234", "feature"),
+                    new GitHubPullRequestClient.Ref("basesha5678", "main")));
+        when(checkRunClient.createCheckRun(
+                anyString(), anyString(), anyString(), anyString(), any()))
+            .thenReturn(new GitHubCheckRunClient.CheckRunResponse(1L, "http://check"));
+        when(prClient.getPullRequestFiles(
+                anyString(), anyString(), anyString(), anyString(), anyInt()))
+            .thenReturn(List.of());
+        when(checkRunClient.getRequiredStatusChecks(
+                anyString(), anyString(), anyString(), anyString(), anyString()))
+            .thenReturn(
+                new GitHubCheckRunClient.RequiredStatusChecks(
+                    List.of("required-build"), List.of()));
+        when(prClient.compareCommits(
+                anyString(), anyString(), anyString(), anyString(), anyString(), anyString()))
+            .thenReturn(new GitHubPullRequestClient.CompareResponse(0, List.of()));
+        when(reviewClient.listReviews(anyString(), anyString(), anyString(), anyString(), anyInt()))
+            .thenReturn(List.of());
+        when(instructionsResolver.resolve(anyString(), anyString(), anyString(), anyLong()))
+            .thenReturn(InstructionsResolver.ResolvedInstructions.EMPTY);
+        when(aiReviewService.review(any(ReviewSession.class), any()))
+            .thenReturn(new ReviewResponse(List.of(), List.of(), null));
+
+        orchestrator.review(manualRequest(""));
+
+        // The session is created with the placeholder, then updated with resolved PR details
+        verify(session).setPrTitle("add new API");
+        verify(session).setCommitSha("headsha1234");
+        var checkRunCaptor =
+            ArgumentCaptor.forClass(GitHubCheckRunClient.CreateCheckRunRequest.class);
+        verify(checkRunClient)
+            .createCheckRun(
+                anyString(), anyString(), eq("owner"), eq("repo"), checkRunCaptor.capture());
+        assertEquals("headsha1234", checkRunCaptor.getValue().headSha());
+      }
     }
 
     @Test
-    void shouldIgnoreBotCommentsThatAreNotTheSummary() {
-      // The bot also posts conversational replies (e.g. answering /review questions); only the
-      // summary heading counts.
-      stubComments(comment("I don't have enough context to answer that.", "thrillhousebot[bot]"));
+    void reviewShouldCatchExceptionWhenFetchingStatusChecks() {
+      try (var mockedStatic = mockStatic(ReviewSession.class)) {
+        var session = mock(ReviewSession.class);
+        session.id = 1L;
+        when(session.getRepository()).thenReturn("owner/repo");
+        when(session.getPrNumber()).thenReturn(7);
+        when(session.getPrTitle()).thenReturn("(manual review)");
+        when(session.getCommitSha()).thenReturn("");
+        when(session.getTimestamp()).thenReturn(java.time.Instant.parse("2025-06-01T12:00:00Z"));
+        mockedStatic
+            .when(() -> ReviewSession.create(anyString(), anyInt(), anyString(), anyString()))
+            .thenReturn(session);
 
-      assertFalse(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
-    }
+        when(authClient.getAuthHeader(123L)).thenReturn("Bearer test");
+        when(prClient.getPullRequest(any(), any(), eq("owner"), eq("repo"), eq(7)))
+            .thenReturn(
+                new GitHubPullRequestClient.PullRequestDetails(
+                    "add new API",
+                    "Adds a new API endpoint",
+                    new GitHubPullRequestClient.Ref("headsha1234", "feature"),
+                    new GitHubPullRequestClient.Ref("basesha5678", "main")));
+        when(prClient.getPullRequestFiles(
+                anyString(), anyString(), anyString(), anyString(), anyInt()))
+            .thenReturn(List.of());
+        when(checkRunClient.getRequiredStatusChecks(
+                anyString(), anyString(), anyString(), anyString(), anyString()))
+            .thenThrow(new RuntimeException("failed"));
+        when(checkRunClient.createCheckRun(
+                anyString(), anyString(), anyString(), anyString(), any()))
+            .thenReturn(new GitHubCheckRunClient.CheckRunResponse(1L, "http://check"));
+        when(prClient.compareCommits(
+                anyString(), anyString(), anyString(), anyString(), anyString(), anyString()))
+            .thenReturn(new GitHubPullRequestClient.CompareResponse(0, List.of()));
+        when(reviewClient.listReviews(anyString(), anyString(), anyString(), anyString(), anyInt()))
+            .thenReturn(List.of());
+        when(instructionsResolver.resolve(anyString(), anyString(), anyString(), anyLong()))
+            .thenReturn(InstructionsResolver.ResolvedInstructions.EMPTY);
+        when(aiReviewService.review(any(ReviewSession.class), any()))
+            .thenReturn(new ReviewResponse(List.of(), List.of(), null));
 
-    @Test
-    void shouldReturnFalseForEmptyComments() {
-      stubComments();
+        orchestrator.review(
+            new ReviewOrchestrator.ReviewRequest(
+                "owner", "repo", 7, "", "(manual review)", "", "", "main", 123L, true));
 
-      assertFalse(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
-    }
-
-    @Test
-    void shouldNotMatchTheHeadingMidComment() {
-      // Only a comment that *starts* with the heading is the bot's summary; an embedded mention is
-      // a
-      // quote in a discussion, not the summary artifact.
-      stubComments(
-          comment("As noted in the " + PrSummaryGenerator.SUMMARY_HEADING, "thrillhousebot[bot]"));
-
-      assertFalse(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
-    }
-
-    @Test
-    void shouldTolerateNullUserAndNullBody() {
-      stubComments(
-          comment("body without user", null),
-          comment(null, "thrillhousebot[bot]"),
-          comment(PrSummaryGenerator.SUMMARY_HEADING, "thrillhousebot[bot]"));
-
-      assertTrue(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
-    }
-
-    @Test
-    void shouldBeBestEffortWhenTheFetchThrows() {
-      when(commentClient.listComments(anyString(), anyString(), anyString(), anyString(), anyInt()))
-          .thenThrow(new RuntimeException("boom"));
-
-      // A thrown fetch falls back to "no summary seen" rather than blocking the review.
-      // (listComments paginates internally and never returns null, so that case can't arise.)
-      assertFalse(orchestrator.botSummaryCommentExists("auth", "owner", "repo", 1));
-      assertTrue(orchestrator.fetchIssueComments("auth", "owner", "repo", 1).isEmpty());
-    }
-  }
-
-  @Nested
-  class ResolveProjectStack {
-
-    @Test
-    void shouldReturnEmptyWhenStackResolverThrows() {
-      when(projectStackResolver.resolve(any(), any(), any(), anyLong()))
-          .thenThrow(new RuntimeException("github down"));
-
-      var stack =
-          orchestrator.resolveProjectStack(
-              new ReviewOrchestrator.ReviewRequest(
-                  "owner", "repo", 1, "sha", "title", "", "base", "main", 123L, false));
-
-      assertEquals("", stack);
+        // The thrown getRequiredStatusChecks must be swallowed: the review still finalizes its
+        // check run and posts a review rather than failing the whole run.
+        verify(checkRunClient)
+            .updateCheckRun(anyString(), anyString(), anyString(), anyString(), anyLong(), any());
+        verify(reviewClient)
+            .createReview(anyString(), anyString(), anyString(), anyString(), anyInt(), any());
+      }
     }
   }
